Move test_add_concurrent from core to fanout
